requirements.txt
#
#######  Pip module requirements  #######
####### for the AGILE DBus Server #######
#
paho-mqtt
pydash
pytz
python-dotenv
flask
requests
queuelib